+++ src/programs/prompt-program.ts
@@ -3,15 +3,15 @@ import EnvironmentService from "../services/environment-service.js";
 import OpenAiChatHelper from "../langchain/open-ai-chat-helper.js";
 import * as toolHelper from "../helpers/agent-tool-helper.js";
 
-class ChatProgram extends ProgramInterface {
+class PromptProgram extends ProgramInterface {
   protected get name(): string {
-    return "chat";
+    return "prompt";
   }
   protected get aliases(): string[] {
-    return ["c"];
+    return ["p"];
   }
   protected get description(): string {
-    return `Runs a chat interface with ChatGPT. various api keys can be set in the environment variables to enable additional features.`;
+    return `Runs a ReAct agent to resolve the users prompt, limited to a single input.`;
   }
   protected get requiredEnvironmentVariables(): string[] {
     return [EnvironmentService.names.OPENAI_API_KEY];
@@ -39,8 +39,8 @@ class ChatProgram extends ProgramInterface {
     console.log(`Running with Tools: [${Object.keys(tools).join(", ")}]`);
 
     // Start chat with the tools
-    return model.chat({ tools });
+    return model.zeroShot({ tools });
   }
 }
 
-export default ChatProgram;
+export default PromptProgram;