+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.1.5] v0.1.5
+
+* Added currency conversion tool to `chat` program.
+* Added alias support for programs.
+* Added `gptcli c` alias for `gptcli chat`.
+* Cleaned up `finnhub` and `currency conversion` tool integrations.
+* Created `prompt` program.
+* Created `zero-shot-react` chat agent for `prompt` program to allow for complex multi-stage reasoning.
+
 ## [0.1.4] v0.1.4
 
 * Updated `chat` program to include stock price checking tool leveraging `finnhub` api.